% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrencode_helpers.R
\name{qrencode_png}
\alias{qrencode_png}
\title{Return a QR encoded string as a base 64 encoded inline png}
\usage{
qrencode_png(to_encode)
}
\arguments{
\item{to_encode}{the data to encode}
}
\description{
Return a QR encoded string as a base 64 encoded inline png
}
\note{
\code{data:image/png;base64,} is prepended to the encoded png
}
\examples{
cat(qrencode_png("http://rud.is/b"))
}

