% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qrencode_helpers.R
\name{qrencode_raster}
\alias{qrencode_raster}
\title{Return a QR encoded string as a raster object}
\usage{
qrencode_raster(to_encode)
}
\description{
Return a QR encoded string as a raster object
}
\examples{
library(raster)
image(qrencode_raster("http://rud.is/b"), asp=1, col=c("black", "white"), axes=FALSE, xlab="", ylab="")
}

