% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qrencode_raw}
\alias{qrencode_raw}
\title{Encodes a string as a QR code}
\usage{
qrencode_raw(to_encode)
}
\arguments{
\item{to_encode}{character string to encode}
}
\description{
Encodes a string as a QR code
}

