% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrencode_helpers.R
\name{qrencode}
\alias{qrencode}
\title{Return a QR encoded string as a matrix}
\usage{
qrencode(to_encode)
}
\arguments{
\item{to_encode}{the data to encode}
}
\description{
Useful if you want to do your own post-processing
}
\examples{
qrencode("http://rud.is/b")
}

