% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrencode_helpers.R
\name{qrencode_raster}
\alias{qrencode_raster}
\title{Return a QR encoded string as a raster object}
\usage{
qrencode_raster(to_encode)
}
\arguments{
\item{to_encode}{the data to encode}
}
\description{
Return a QR encoded string as a raster object
}
\examples{
library(raster)
old_mar <- par()$mar
par(mar=c(0,0,0,0))
image(qrencode_raster("http://rud.is/b"), asp=1, col=c("white", "black"),
                                          axes=FALSE, xlab="", ylab="")
par(mar=old_mar)
}

